item2/requirements.txt
absl-py==2.1.0
asttokens==2.4.1
astunparse==1.6.3
autopep8==2.0.4
blinker==1.7.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.12.4
Flask==3.0.2
flatbuffers==23.5.26
fonttools==4.49.0
fsspec==2023.9.2
gast==0.4.0
google-auth==2.28.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.60.1
gunicorn==21.2.0
h5py==3.10.0
idna==3.4
imageio==2.34.0
ipykernel==6.29.2
ipython==8.22.1
isort==5.13.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.7.1
keras==2.13.1
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
Markdown==3.5.2
MarkupSafe==2.1.3
matplotlib==3.8.3
matplotlib-inline==0.1.6
ml-dtypes==0.2.0
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pandas==2.2.0
parso==0.8.3
Pillow==10.0.1
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.11.1
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
stack-data==0.6.3
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
threadpoolctl==3.3.0
tifffile==2024.2.12
tomli==2.0.1
torch==2.1.0
torch-tb-profiler==0.4.3
torchaudio==2.1.0
torchvision==0.16.0
tornado==6.4
traitlets==5.14.1
typing_extensions==4.5.0
tzdata==2024.1
urllib3==2.0.6
wcwidth==0.2.13
Werkzeug==3.0.1
wrapt==1.14.1
